refactor: rename file name to kebab-case
