--- src/shared/utils/dateFormat/index.ts
@@ -1,1 +1,0 @@
-export * from "./dateFormat";
