+++ src/shared/utils/define-value-created-at/define-value-created-at.ts
@@ -1,4 +1,4 @@
-import { dateFormat } from "shared/utils/dateFormat";
+import { dateFormat } from "shared/utils/date-format";
 
 export const defineValueCreatedAt = (value: string | undefined) =>
 	value ? dateFormat(value) : "";
